refactor: rename app to server
